Fixing tests.
Some tests were flaky because of the order of instances emitted by `InstanceCollector` was not deterministic (Used `map.keyset()` on a `ConcurrentHashMap`).
Provided a way to manipulate the map instance used, although error prone (since the map has to be concurrent), at this point, this was the easiest way out. The actual RxNetty module load balancer does not use this collector.

The RxNetty tests were broken with the new snapshot, the `ConnectionProvider` requires an explicit `start()`